Change activation/deactivation of wrestlers
